Service/cleanup

+++ src/main/java/smart/housing/controllers/BudgetManagementController.java
@@ -18,8 +18,8 @@ import smart.housing.services.BudgetManagementService;
 import smart.housing.services.BudgetManagementServiceImplementation;
 import smart.housing.services.UserManagementService;
 import smart.housing.ui.*;
-import javafx.collections.ListChangeListener;
-
+import java.io.IOException;
+import java.net.URISyntaxException;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -74,17 +74,6 @@ public class BudgetManagementController extends SmartHousingController {
      */
     public void initialize() {
         setBackgroundImage();
-
-        // Set up listener for multiple selections
-        debtors.getCheckModel().getCheckedItems().addListener((ListChangeListener<? super User>) change -> {
-            while (change.next()) {
-                if (change.wasAdded()) {
-                    System.out.println("Selected Debtors: " + change.getAddedSubList());
-                } else if (change.wasRemoved()) {
-                    System.out.println("Deselected Debtors: " + change.getRemoved());
-                }
-            }
-        });
 
         // Set up listener for selection in debtsOverview
         debtsOverview.getSelectionModel().selectedItemProperty().addListener((obs, oldSelection, newSelection) -> {
@@ -148,7 +137,6 @@ public class BudgetManagementController extends SmartHousingController {
         addExpenseButtonClicked();
     }
 
-
     private void addExpenseButtonClicked() {
         try {
             String product = productNameField.getText();
@@ -210,7 +198,12 @@ public class BudgetManagementController extends SmartHousingController {
     }
 
     private void removeItemFromList(Expense expense) {
-        BUDGET_SERVICE.delete(expense);
+        new ConfirmDialog(
+            "Delete selected expense?", "Delete", "Keep"
+        ).showAndWait().ifPresent(aBoolean -> {
+            if(aBoolean)
+                BUDGET_SERVICE.delete(expense);
+        });
         loadExpenseList();
         loadDebtsOverviewList();
     }
@@ -261,41 +254,30 @@ public class BudgetManagementController extends SmartHousingController {
         DebtOverview selectedDebt = debtsOverview.getSelectionModel().getSelectedItem();
 
         if (selectedDebt != null) {
-            try {
-                User creditor = selectedDebt.creditor();
-                User debtor = selectedDebt.debtor();
-                double openDebt = selectedDebt.getAmount();
-
-                // Add an expense with the Name "Settled Debt"
-                BUDGET_SERVICE.create(new Expense(Set.of(debtor), creditor, "Settled Debt", (openDebt*-1)));
-
-                // Display success message
-                buttonDisplay(true, settleDebtButton);
-
-            } catch (Exception e) {
-                // Display error message
-                buttonDisplay(false, settleDebtButton);
-                e.printStackTrace();
-            } finally {
-                clearExpenses();
-                loadExpenseList();
-                loadDebtsOverviewList();
-            }
-        } else {
-            // No row selected, display error message
-            buttonDisplay(false, settleDebtButton);
+            User contrary = selectedDebt.getCreditor();
+            List<Expense> expenseList = BUDGET_SERVICE.getAllExpenses()
+                .stream()
+                .filter(expense ->
+                    (expense.getCreditor().equals(contrary) && expense.getDebtors().contains(USER_SERVICE.getServiceUser())) ||
+                    (expense.getCreditor().equals(USER_SERVICE.getServiceUser()) && expense.getDebtors().contains(contrary))
+                ).toList();
+            for(Expense expense : expenseList)
+                BUDGET_SERVICE.delete(expense);
+            update();
         }
     }
 
     private void sendEmail() {
         // Get the selected row from the debtsOverview table
         DebtOverview selectedDebt = debtsOverview.getSelectionModel().getSelectedItem();
-        String senderFirstName = USER_SERVICE.getServiceUser().getFirstName();
-        String senderLastName = USER_SERVICE.getServiceUser().getLastName();
 
         if (selectedDebt != null) {
-            BUDGET_SERVICE.sendReminderMail(selectedDebt, senderFirstName, senderLastName);
-            buttonDisplay(true, emailButton);
+            try {
+                BUDGET_SERVICE.sendReminderEmail(USER_SERVICE.getServiceUser(), selectedDebt);
+                buttonDisplay(true, emailButton);
+            } catch (IOException | URISyntaxException e) {
+                buttonDisplay(false, emailButton);
+            }
         } else {
             buttonDisplay(false, emailButton);
         }
